fix: page.tsx에서   prefetch 하는 패턴으로 변경
